move Headers/Message/Request/Response into mitmproxy.http
