+++ examples/addons/http-reply-from-proxy.py
@@ -4,7 +4,7 @@ from mitmproxy import http
 
 def request(flow: http.HTTPFlow) -> None:
     if flow.request.pretty_url == "http://example.com/path":
-        flow.response = http.HTTPResponse.make(
+        flow.response = http.Response.make(
             200,  # (optional) status code
             b"Hello World",  # (optional) content
             {"Content-Type": "text/html"}  # (optional) headers
