+++ mitmproxy/addons/proxyauth.py
@@ -81,16 +81,16 @@ class ProxyAuth:
         else:
             return 'Authorization'
 
-    def auth_required_response(self) -> http.HTTPResponse:
+    def auth_required_response(self) -> http.Response:
         if self.is_proxy_auth():
             return http.make_error_response(
                 status_codes.PROXY_AUTH_REQUIRED,
-                headers=mitmproxy.net.http.Headers(Proxy_Authenticate=f'Basic realm="{REALM}"'),
+                headers=mitmproxy.http.Headers(Proxy_Authenticate=f'Basic realm="{REALM}"'),
             )
         else:
             return http.make_error_response(
                 status_codes.UNAUTHORIZED,
-                headers=mitmproxy.net.http.Headers(WWW_Authenticate=f'Basic realm="{REALM}"'),
+                headers=mitmproxy.http.Headers(WWW_Authenticate=f'Basic realm="{REALM}"'),
             )
 
     def check(self, f: http.HTTPFlow) -> Optional[Tuple[str, str]]:
