+++ mitmproxy/http.py
@@ -1,14 +1,1108 @@
 import html
+import re
 import time
-from typing import Optional, Tuple
+import urllib.parse
+from dataclasses import dataclass
+from dataclasses import fields
+from email.utils import formatdate, mktime_tz, parsedate_tz
+from typing import Callable, cast
+from typing import Dict
+from typing import Iterable
+from typing import Mapping
+from typing import Optional
+from typing import Tuple
+from typing import Union
 
+from mitmproxy.net.http import url
 from mitmproxy import flow
 from mitmproxy import version
-from mitmproxy.net import http
+from mitmproxy.coretypes import multidict
+from mitmproxy.coretypes import serializable
+from mitmproxy.net import http, encoding
+from mitmproxy.net.http import cookies, multipart
+from mitmproxy.net.http import status_codes
+from mitmproxy.net.http.headers import assemble_content_type, parse_content_type
 from mitmproxy.proxy import context
+from mitmproxy.utils import human
+from mitmproxy.utils import strutils
+from mitmproxy.utils import typecheck
+from mitmproxy.utils.strutils import always_bytes
+from mitmproxy.utils.strutils import always_str
 
-HTTPRequest = http.Request
-HTTPResponse = http.Response
+
+# While headers _should_ be ASCII, it's not uncommon for certain headers to be utf-8 encoded.
+def _native(x):
+    return x.decode("utf-8", "surrogateescape")
+
+
+def _always_bytes(x):
+    return strutils.always_bytes(x, "utf-8", "surrogateescape")
+
+
+class Headers(multidict.MultiDict):
+    """
+    Header class which allows both convenient access to individual headers as well as
+    direct access to the underlying raw data. Provides a full dictionary interface.
+
+    Example:
+
+    .. code-block:: python
+
+        # Create headers with keyword arguments
+        >>> h = Headers(host="example.com", content_type="application/xml")
+
+        # Headers mostly behave like a normal dict.
+        >>> h["Host"]
+        "example.com"
+
+        # HTTP Headers are case insensitive
+        >>> h["host"]
+        "example.com"
+
+        # Headers can also be created from a list of raw (header_name, header_value) byte tuples
+        >>> h = Headers([
+            (b"Host",b"example.com"),
+            (b"Accept",b"text/html"),
+            (b"accept",b"application/xml")
+        ])
+
+        # Multiple headers are folded into a single header as per RFC7230
+        >>> h["Accept"]
+        "text/html, application/xml"
+
+        # Setting a header removes all existing headers with the same name.
+        >>> h["Accept"] = "application/text"
+        >>> h["Accept"]
+        "application/text"
+
+        # bytes(h) returns a HTTP1 header block.
+        >>> print(bytes(h))
+        Host: example.com
+        Accept: application/text
+
+        # For full control, the raw header fields can be accessed
+        >>> h.fields
+
+    Caveats:
+        For use with the "Set-Cookie" header, see :py:meth:`get_all`.
+    """
+
+    def __init__(self, fields=(), **headers):
+        """
+        Args:
+            fields: (optional) list of ``(name, value)`` header byte tuples,
+                e.g. ``[(b"Host", b"example.com")]``. All names and values must be bytes.
+            **headers: Additional headers to set. Will overwrite existing values from `fields`.
+                For convenience, underscores in header names will be transformed to dashes -
+                this behaviour does not extend to other methods.
+                If ``**headers`` contains multiple keys that have equal ``.lower()`` s,
+                the behavior is undefined.
+        """
+        super().__init__(fields)
+
+        for key, value in self.fields:
+            if not isinstance(key, bytes) or not isinstance(value, bytes):
+                raise TypeError("Header fields must be bytes.")
+
+        # content_type -> content-type
+        headers = {
+            _always_bytes(name).replace(b"_", b"-"): _always_bytes(value)
+            for name, value in headers.items()
+        }
+        self.update(headers)
+
+    @staticmethod
+    def _reduce_values(values):
+        # Headers can be folded
+        return ", ".join(values)
+
+    @staticmethod
+    def _kconv(key):
+        # Headers are case-insensitive
+        return key.lower()
+
+    def __bytes__(self):
+        if self.fields:
+            return b"\r\n".join(b": ".join(field) for field in self.fields) + b"\r\n"
+        else:
+            return b""
+
+    def __delitem__(self, key):
+        key = _always_bytes(key)
+        super().__delitem__(key)
+
+    def __iter__(self):
+        for x in super().__iter__():
+            yield _native(x)
+
+    def get_all(self, name):
+        """
+        Like :py:meth:`get`, but does not fold multiple headers into a single one.
+        This is useful for Set-Cookie headers, which do not support folding.
+        See also: https://tools.ietf.org/html/rfc7230#section-3.2.2
+        """
+        name = _always_bytes(name)
+        return [
+            _native(x) for x in
+            super().get_all(name)
+        ]
+
+    def set_all(self, name, values):
+        """
+        Explicitly set multiple headers for the given key.
+        See: :py:meth:`get_all`
+        """
+        name = _always_bytes(name)
+        values = [_always_bytes(x) for x in values]
+        return super().set_all(name, values)
+
+    def insert(self, index, key, value):
+        key = _always_bytes(key)
+        value = _always_bytes(value)
+        super().insert(index, key, value)
+
+    def items(self, multi=False):
+        if multi:
+            return (
+                (_native(k), _native(v))
+                for k, v in self.fields
+            )
+        else:
+            return super().items()
+
+
+@dataclass
+class MessageData(serializable.Serializable):
+    http_version: bytes
+    headers: Headers
+    content: Optional[bytes]
+    trailers: Optional[Headers]
+    timestamp_start: float
+    timestamp_end: Optional[float]
+
+    # noinspection PyUnreachableCode
+    if __debug__:
+        def __post_init__(self):
+            for field in fields(self):
+                val = getattr(self, field.name)
+                typecheck.check_option_type(field.name, val, field.type)
+
+    def set_state(self, state):
+        for k, v in state.items():
+            if k in ("headers", "trailers") and v is not None:
+                v = Headers.from_state(v)
+            setattr(self, k, v)
+
+    def get_state(self):
+        state = vars(self).copy()
+        state["headers"] = state["headers"].get_state()
+        if state["trailers"] is not None:
+            state["trailers"] = state["trailers"].get_state()
+        return state
+
+    @classmethod
+    def from_state(cls, state):
+        state["headers"] = Headers.from_state(state["headers"])
+        if state["trailers"] is not None:
+            state["trailers"] = Headers.from_state(state["trailers"])
+        return cls(**state)
+
+
+@dataclass
+class RequestData(MessageData):
+    host: str
+    port: int
+    method: bytes
+    scheme: bytes
+    authority: bytes
+    path: bytes
+
+
+@dataclass
+class ResponseData(MessageData):
+    status_code: int
+    reason: bytes
+
+
+class Message(serializable.Serializable):
+    @classmethod
+    def from_state(cls, state):
+        return cls(**state)
+
+    def get_state(self):
+        return self.data.get_state()
+
+    def set_state(self, state):
+        self.data.set_state(state)
+
+    data: MessageData
+    stream: Union[Callable, bool] = False
+
+    @property
+    def http_version(self) -> str:
+        """
+        Version string, e.g. "HTTP/1.1"
+        """
+        return self.data.http_version.decode("utf-8", "surrogateescape")
+
+    @http_version.setter
+    def http_version(self, http_version: Union[str, bytes]) -> None:
+        self.data.http_version = strutils.always_bytes(http_version, "utf-8", "surrogateescape")
+
+    @property
+    def is_http10(self) -> bool:
+        return self.data.http_version == b"HTTP/1.0"
+
+    @property
+    def is_http11(self) -> bool:
+        return self.data.http_version == b"HTTP/1.1"
+
+    @property
+    def is_http2(self) -> bool:
+        return self.data.http_version == b"HTTP/2.0"
+
+    @property
+    def headers(self) -> Headers:
+        """
+        The HTTP headers.
+        """
+        return self.data.headers
+
+    @headers.setter
+    def headers(self, h: Headers) -> None:
+        self.data.headers = h
+
+    @property
+    def trailers(self) -> Optional[Headers]:
+        """
+        The HTTP trailers.
+        """
+        return self.data.trailers
+
+    @trailers.setter
+    def trailers(self, h: Optional[Headers]) -> None:
+        self.data.trailers = h
+
+    @property
+    def raw_content(self) -> Optional[bytes]:
+        """
+        The raw (potentially compressed) HTTP message body as bytes.
+
+        See also: :py:attr:`content`, :py:class:`text`
+        """
+        return self.data.content
+
+    @raw_content.setter
+    def raw_content(self, content: Optional[bytes]) -> None:
+        self.data.content = content
+
+    def get_content(self, strict: bool = True) -> Optional[bytes]:
+        """
+        The uncompressed HTTP message body as bytes.
+
+        Raises:
+            ValueError, when the HTTP content-encoding is invalid and strict is True.
+
+        See also: :py:class:`raw_content`, :py:attr:`text`
+        """
+        if self.raw_content is None:
+            return None
+        ce = self.headers.get("content-encoding")
+        if ce:
+            try:
+                content = encoding.decode(self.raw_content, ce)
+                # A client may illegally specify a byte -> str encoding here (e.g. utf8)
+                if isinstance(content, str):
+                    raise ValueError(f"Invalid Content-Encoding: {ce}")
+                return content
+            except ValueError:
+                if strict:
+                    raise
+                return self.raw_content
+        else:
+            return self.raw_content
+
+    def set_content(self, value: Optional[bytes]) -> None:
+        if value is None:
+            self.raw_content = None
+            return
+        if not isinstance(value, bytes):
+            raise TypeError(
+                f"Message content must be bytes, not {type(value).__name__}. "
+                "Please use .text if you want to assign a str."
+            )
+        ce = self.headers.get("content-encoding")
+        try:
+            self.raw_content = encoding.encode(value, ce or "identity")
+        except ValueError:
+            # So we have an invalid content-encoding?
+            # Let's remove it!
+            del self.headers["content-encoding"]
+            self.raw_content = value
+        self.headers["content-length"] = str(len(self.raw_content))
+
+    content = property(get_content, set_content)
+
+    @property
+    def timestamp_start(self) -> float:
+        """
+        First byte timestamp
+        """
+        return self.data.timestamp_start
+
+    @timestamp_start.setter
+    def timestamp_start(self, timestamp_start: float) -> None:
+        self.data.timestamp_start = timestamp_start
+
+    @property
+    def timestamp_end(self) -> Optional[float]:
+        """
+        Last byte timestamp
+        """
+        return self.data.timestamp_end
+
+    @timestamp_end.setter
+    def timestamp_end(self, timestamp_end: Optional[float]):
+        self.data.timestamp_end = timestamp_end
+
+    def _get_content_type_charset(self) -> Optional[str]:
+        ct = parse_content_type(self.headers.get("content-type", ""))
+        if ct:
+            return ct[2].get("charset")
+        return None
+
+    def _guess_encoding(self, content: bytes = b"") -> str:
+        enc = self._get_content_type_charset()
+        if not enc:
+            if "json" in self.headers.get("content-type", ""):
+                enc = "utf8"
+        if not enc:
+            meta_charset = re.search(rb"""<meta[^>]+charset=['"]?([^'">]+)""", content)
+            if meta_charset:
+                enc = meta_charset.group(1).decode("ascii", "ignore")
+        if not enc:
+            if "text/css" in self.headers.get("content-type", ""):
+                # @charset rule must be the very first thing.
+                css_charset = re.match(rb"""@charset "([^"]+)";""", content)
+                if css_charset:
+                    enc = css_charset.group(1).decode("ascii", "ignore")
+        if not enc:
+            enc = "latin-1"
+        # Use GB 18030 as the superset of GB2312 and GBK to fix common encoding problems on Chinese websites.
+        if enc.lower() in ("gb2312", "gbk"):
+            enc = "gb18030"
+
+        return enc
+
+    def get_text(self, strict: bool = True) -> Optional[str]:
+        """
+        The uncompressed and decoded HTTP message body as text.
+
+        Raises:
+            ValueError, when either content-encoding or charset is invalid and strict is True.
+
+        See also: :py:attr:`content`, :py:class:`raw_content`
+        """
+        content = self.get_content(strict)
+        if content is None:
+            return None
+        enc = self._guess_encoding(content)
+        try:
+            return cast(str, encoding.decode(content, enc))
+        except ValueError:
+            if strict:
+                raise
+            return content.decode("utf8", "surrogateescape")
+
+    def set_text(self, text: Optional[str]) -> None:
+        if text is None:
+            self.content = None
+            return
+        enc = self._guess_encoding()
+
+        try:
+            self.content = encoding.encode(text, enc)
+        except ValueError:
+            # Fall back to UTF-8 and update the content-type header.
+            ct = parse_content_type(self.headers.get("content-type", "")) or ("text", "plain", {})
+            ct[2]["charset"] = "utf-8"
+            self.headers["content-type"] = assemble_content_type(*ct)
+            enc = "utf8"
+            self.content = text.encode(enc, "surrogateescape")
+
+    text = property(get_text, set_text)
+
+    def decode(self, strict: bool = True) -> None:
+        """
+        Decodes body based on the current Content-Encoding header, then
+        removes the header. If there is no Content-Encoding header, no
+        action is taken.
+
+        Raises:
+            ValueError, when the content-encoding is invalid and strict is True.
+        """
+        decoded = self.get_content(strict)
+        self.headers.pop("content-encoding", None)
+        self.content = decoded
+
+    def encode(self, e: str) -> None:
+        """
+        Encodes body with the encoding e, where e is "gzip", "deflate", "identity", "br", or "zstd".
+        Any existing content-encodings are overwritten,
+        the content is not decoded beforehand.
+
+        Raises:
+            ValueError, when the specified content-encoding is invalid.
+        """
+        self.headers["content-encoding"] = e
+        self.content = self.raw_content
+        if "content-encoding" not in self.headers:
+            raise ValueError("Invalid content encoding {}".format(repr(e)))
+
+
+class Request(Message):
+    """
+    An HTTP request.
+    """
+    data: RequestData
+
+    def __init__(
+        self,
+        host: str,
+        port: int,
+        method: bytes,
+        scheme: bytes,
+        authority: bytes,
+        path: bytes,
+        http_version: bytes,
+        headers: Union[Headers, Tuple[Tuple[bytes, bytes], ...]],
+        content: Optional[bytes],
+        trailers: Union[None, Headers, Tuple[Tuple[bytes, bytes], ...]],
+        timestamp_start: float,
+        timestamp_end: Optional[float],
+    ):
+        # auto-convert invalid types to retain compatibility with older code.
+        if isinstance(host, bytes):
+            host = host.decode("idna", "strict")
+        if isinstance(method, str):
+            method = method.encode("ascii", "strict")
+        if isinstance(scheme, str):
+            scheme = scheme.encode("ascii", "strict")
+        if isinstance(authority, str):
+            authority = authority.encode("ascii", "strict")
+        if isinstance(path, str):
+            path = path.encode("ascii", "strict")
+        if isinstance(http_version, str):
+            http_version = http_version.encode("ascii", "strict")
+
+        if isinstance(content, str):
+            raise ValueError(f"Content must be bytes, not {type(content).__name__}")
+        if not isinstance(headers, Headers):
+            headers = Headers(headers)
+        if trailers is not None and not isinstance(trailers, Headers):
+            trailers = Headers(trailers)
+
+        self.data = RequestData(
+            host=host,
+            port=port,
+            method=method,
+            scheme=scheme,
+            authority=authority,
+            path=path,
+            http_version=http_version,
+            headers=headers,
+            content=content,
+            trailers=trailers,
+            timestamp_start=timestamp_start,
+            timestamp_end=timestamp_end,
+        )
+
+    def __repr__(self) -> str:
+        if self.host and self.port:
+            hostport = f"{self.host}:{self.port}"
+        else:
+            hostport = ""
+        path = self.path or ""
+        return f"Request({self.method} {hostport}{path})"
+
+    @classmethod
+    def make(
+        cls,
+        method: str,
+        url: str,
+        content: Union[bytes, str] = "",
+        headers: Union[Headers, Dict[Union[str, bytes], Union[str, bytes]], Iterable[Tuple[bytes, bytes]]] = ()
+    ) -> "Request":
+        """
+        Simplified API for creating request objects.
+        """
+        # Headers can be list or dict, we differentiate here.
+        if isinstance(headers, Headers):
+            pass
+        elif isinstance(headers, dict):
+            headers = Headers(
+                (always_bytes(k, "utf-8", "surrogateescape"),
+                 always_bytes(v, "utf-8", "surrogateescape"))
+                for k, v in headers.items()
+            )
+        elif isinstance(headers, Iterable):
+            headers = Headers(headers)
+        else:
+            raise TypeError("Expected headers to be an iterable or dict, but is {}.".format(
+                type(headers).__name__
+            ))
+
+        req = cls(
+            "",
+            0,
+            method.encode("utf-8", "surrogateescape"),
+            b"",
+            b"",
+            b"",
+            b"HTTP/1.1",
+            headers,
+            b"",
+            None,
+            time.time(),
+            time.time(),
+        )
+
+        req.url = url
+        # Assign this manually to update the content-length header.
+        if isinstance(content, bytes):
+            req.content = content
+        elif isinstance(content, str):
+            req.text = content
+        else:
+            raise TypeError(f"Expected content to be str or bytes, but is {type(content).__name__}.")
+
+        return req
+
+    @property
+    def first_line_format(self) -> str:
+        """
+        HTTP request form as defined in `RFC7230 <https://tools.ietf.org/html/rfc7230#section-5.3>`_.
+
+        origin-form and asterisk-form are subsumed as "relative".
+        """
+        if self.method == "CONNECT":
+            return "authority"
+        elif self.authority:
+            return "absolute"
+        else:
+            return "relative"
+
+    @property
+    def method(self) -> str:
+        """
+        HTTP request method, e.g. "GET".
+        """
+        return self.data.method.decode("utf-8", "surrogateescape").upper()
+
+    @method.setter
+    def method(self, val: Union[str, bytes]) -> None:
+        self.data.method = always_bytes(val, "utf-8", "surrogateescape")
+
+    @property
+    def scheme(self) -> str:
+        """
+        HTTP request scheme, which should be "http" or "https".
+        """
+        return self.data.scheme.decode("utf-8", "surrogateescape")
+
+    @scheme.setter
+    def scheme(self, val: Union[str, bytes]) -> None:
+        self.data.scheme = always_bytes(val, "utf-8", "surrogateescape")
+
+    @property
+    def authority(self) -> str:
+        """
+        HTTP request authority.
+
+        For HTTP/1, this is the authority portion of the request target
+        (in either absolute-form or authority-form)
+
+        For HTTP/2, this is the :authority pseudo header.
+        """
+        try:
+            return self.data.authority.decode("idna")
+        except UnicodeError:
+            return self.data.authority.decode("utf8", "surrogateescape")
+
+    @authority.setter
+    def authority(self, val: Union[str, bytes]) -> None:
+        if isinstance(val, str):
+            try:
+                val = val.encode("idna", "strict")
+            except UnicodeError:
+                val = val.encode("utf8", "surrogateescape")  # type: ignore
+        self.data.authority = val
+
+    @property
+    def host(self) -> str:
+        """
+        Target host. This may be parsed from the raw request
+        (e.g. from a ``GET http://example.com/ HTTP/1.1`` request line)
+        or inferred from the proxy mode (e.g. an IP in transparent mode).
+
+        Setting the host attribute also updates the host header and authority information, if present.
+        """
+        return self.data.host
+
+    @host.setter
+    def host(self, val: Union[str, bytes]) -> None:
+        self.data.host = always_str(val, "idna", "strict")
+
+        # Update host header
+        if "Host" in self.data.headers:
+            self.data.headers["Host"] = val
+        # Update authority
+        if self.data.authority:
+            self.authority = url.hostport(self.scheme, self.host, self.port)
+
+    @property
+    def host_header(self) -> Optional[str]:
+        """
+        The request's host/authority header.
+
+        This property maps to either ``request.headers["Host"]`` or
+        ``request.authority``, depending on whether it's HTTP/1.x or HTTP/2.0.
+        """
+        if self.is_http2:
+            return self.authority or self.data.headers.get("Host", None)
+        else:
+            return self.data.headers.get("Host", None)
+
+    @host_header.setter
+    def host_header(self, val: Union[None, str, bytes]) -> None:
+        if val is None:
+            if self.is_http2:
+                self.data.authority = b""
+            self.headers.pop("Host", None)
+        else:
+            if self.is_http2:
+                self.authority = val  # type: ignore
+            if not self.is_http2 or "Host" in self.headers:
+                # For h2, we only overwrite, but not create, as :authority is the h2 host header.
+                self.headers["Host"] = val
+
+    @property
+    def port(self) -> int:
+        """
+        Target port
+        """
+        return self.data.port
+
+    @port.setter
+    def port(self, port: int) -> None:
+        self.data.port = port
+
+    @property
+    def path(self) -> str:
+        """
+        HTTP request path, e.g. "/index.html".
+        Usually starts with a slash, except for OPTIONS requests, which may just be "*".
+        """
+        return self.data.path.decode("utf-8", "surrogateescape")
+
+    @path.setter
+    def path(self, val: Union[str, bytes]) -> None:
+        self.data.path = always_bytes(val, "utf-8", "surrogateescape")
+
+    @property
+    def url(self) -> str:
+        """
+        The URL string, constructed from the request's URL components.
+        """
+        if self.first_line_format == "authority":
+            return f"{self.host}:{self.port}"
+        return url.unparse(self.scheme, self.host, self.port, self.path)
+
+    @url.setter
+    def url(self, val: Union[str, bytes]) -> None:
+        val = always_str(val, "utf-8", "surrogateescape")
+        self.scheme, self.host, self.port, self.path = url.parse(val)
+
+    @property
+    def pretty_host(self) -> str:
+        """
+        Similar to :py:attr:`host`, but using the host/:authority header as an additional (preferred) data source.
+        This is useful in transparent mode where :py:attr:`host` is only an IP address,
+        but may not reflect the actual destination as the Host header could be spoofed.
+        """
+        authority = self.host_header
+        if authority:
+            return url.parse_authority(authority, check=False)[0]
+        else:
+            return self.host
+
+    @property
+    def pretty_url(self) -> str:
+        """
+        Like :py:attr:`url`, but using :py:attr:`pretty_host` instead of :py:attr:`host`.
+        """
+        if self.first_line_format == "authority":
+            return self.authority
+
+        host_header = self.host_header
+        if not host_header:
+            return self.url
+
+        pretty_host, pretty_port = url.parse_authority(host_header, check=False)
+        pretty_port = pretty_port or url.default_port(self.scheme) or 443
+
+        return url.unparse(self.scheme, pretty_host, pretty_port, self.path)
+
+    def _get_query(self):
+        query = urllib.parse.urlparse(self.url).query
+        return tuple(url.decode(query))
+
+    def _set_query(self, query_data):
+        query = url.encode(query_data)
+        _, _, path, params, _, fragment = urllib.parse.urlparse(self.url)
+        self.path = urllib.parse.urlunparse(["", "", path, params, query, fragment])
+
+    @property
+    def query(self) -> multidict.MultiDictView:
+        """
+        The request query string as an :py:class:`~mitmproxy.net.multidict.MultiDictView` object.
+        """
+        return multidict.MultiDictView(
+            self._get_query,
+            self._set_query
+        )
+
+    @query.setter
+    def query(self, value):
+        self._set_query(value)
+
+    def _get_cookies(self):
+        h = self.headers.get_all("Cookie")
+        return tuple(cookies.parse_cookie_headers(h))
+
+    def _set_cookies(self, value):
+        self.headers["cookie"] = cookies.format_cookie_header(value)
+
+    @property
+    def cookies(self) -> multidict.MultiDictView:
+        """
+        The request cookies.
+
+        An empty :py:class:`~mitmproxy.net.multidict.MultiDictView` object if the cookie monster ate them all.
+        """
+        return multidict.MultiDictView(
+            self._get_cookies,
+            self._set_cookies
+        )
+
+    @cookies.setter
+    def cookies(self, value):
+        self._set_cookies(value)
+
+    @property
+    def path_components(self):
+        """
+        The URL's path components as a tuple of strings.
+        Components are unquoted.
+        """
+        path = urllib.parse.urlparse(self.url).path
+        # This needs to be a tuple so that it's immutable.
+        # Otherwise, this would fail silently:
+        #   request.path_components.append("foo")
+        return tuple(url.unquote(i) for i in path.split("/") if i)
+
+    @path_components.setter
+    def path_components(self, components):
+        components = map(lambda x: url.quote(x, safe=""), components)
+        path = "/" + "/".join(components)
+        _, _, _, params, query, fragment = urllib.parse.urlparse(self.url)
+        self.path = urllib.parse.urlunparse(["", "", path, params, query, fragment])
+
+    def anticache(self) -> None:
+        """
+        Modifies this request to remove headers that might produce a cached
+        response. That is, we remove ETags and If-Modified-Since headers.
+        """
+        delheaders = [
+            "if-modified-since",
+            "if-none-match",
+        ]
+        for i in delheaders:
+            self.headers.pop(i, None)
+
+    def anticomp(self) -> None:
+        """
+        Modifies this request to remove headers that will compress the
+        resource's data.
+        """
+        self.headers["accept-encoding"] = "identity"
+
+    def constrain_encoding(self) -> None:
+        """
+        Limits the permissible Accept-Encoding values, based on what we can
+        decode appropriately.
+        """
+        accept_encoding = self.headers.get("accept-encoding")
+        if accept_encoding:
+            self.headers["accept-encoding"] = (
+                ', '.join(
+                    e
+                    for e in {"gzip", "identity", "deflate", "br", "zstd"}
+                    if e in accept_encoding
+                )
+            )
+
+    def _get_urlencoded_form(self):
+        is_valid_content_type = "application/x-www-form-urlencoded" in self.headers.get("content-type", "").lower()
+        if is_valid_content_type:
+            return tuple(url.decode(self.get_text(strict=False)))
+        return ()
+
+    def _set_urlencoded_form(self, form_data):
+        """
+        Sets the body to the URL-encoded form data, and adds the appropriate content-type header.
+        This will overwrite the existing content if there is one.
+        """
+        self.headers["content-type"] = "application/x-www-form-urlencoded"
+        self.content = url.encode(form_data, self.get_text(strict=False)).encode()
+
+    @property
+    def urlencoded_form(self):
+        """
+        The URL-encoded form data as an :py:class:`~mitmproxy.net.multidict.MultiDictView` object.
+        An empty multidict.MultiDictView if the content-type indicates non-form data
+        or the content could not be parsed.
+
+        Starting with mitmproxy 1.0, key and value are strings.
+        """
+        return multidict.MultiDictView(
+            self._get_urlencoded_form,
+            self._set_urlencoded_form
+        )
+
+    @urlencoded_form.setter
+    def urlencoded_form(self, value):
+        self._set_urlencoded_form(value)
+
+    def _get_multipart_form(self):
+        is_valid_content_type = "multipart/form-data" in self.headers.get("content-type", "").lower()
+        if is_valid_content_type:
+            try:
+                return multipart.decode(self.headers.get("content-type"), self.content)
+            except ValueError:
+                pass
+        return ()
+
+    def _set_multipart_form(self, value):
+        self.content = multipart.encode(self.headers, value)
+        self.headers["content-type"] = "multipart/form-data"
+
+    @property
+    def multipart_form(self):
+        """
+        The multipart form data as an :py:class:`~mitmproxy.net.multidict.MultiDictView` object.
+        An empty multidict.MultiDictView if the content-type indicates non-form data
+        or the content could not be parsed.
+
+        Key and value are bytes.
+        """
+        return multidict.MultiDictView(
+            self._get_multipart_form,
+            self._set_multipart_form
+        )
+
+    @multipart_form.setter
+    def multipart_form(self, value):
+        self._set_multipart_form(value)
+
+
+class Response(Message):
+    """
+    An HTTP response.
+    """
+    data: ResponseData
+
+    def __init__(
+        self,
+        http_version: bytes,
+        status_code: int,
+        reason: bytes,
+        headers: Union[Headers, Tuple[Tuple[bytes, bytes], ...]],
+        content: Optional[bytes],
+        trailers: Union[None, Headers, Tuple[Tuple[bytes, bytes], ...]],
+        timestamp_start: float,
+        timestamp_end: Optional[float],
+    ):
+        # auto-convert invalid types to retain compatibility with older code.
+        if isinstance(http_version, str):
+            http_version = http_version.encode("ascii", "strict")
+        if isinstance(reason, str):
+            reason = reason.encode("ascii", "strict")
+
+        if isinstance(content, str):
+            raise ValueError("Content must be bytes, not {}".format(type(content).__name__))
+        if not isinstance(headers, Headers):
+            headers = Headers(headers)
+        if trailers is not None and not isinstance(trailers, Headers):
+            trailers = Headers(trailers)
+
+        self.data = ResponseData(
+            http_version=http_version,
+            status_code=status_code,
+            reason=reason,
+            headers=headers,
+            content=content,
+            trailers=trailers,
+            timestamp_start=timestamp_start,
+            timestamp_end=timestamp_end,
+        )
+
+    def __repr__(self) -> str:
+        if self.raw_content:
+            ct = self.headers.get("content-type", "unknown content type")
+            size = human.pretty_size(len(self.raw_content))
+            details = f"{ct}, {size}"
+        else:
+            details = "no content"
+        return f"Response({self.status_code}, {details})"
+
+    @classmethod
+    def make(
+        cls,
+        status_code: int = 200,
+        content: Union[bytes, str] = b"",
+        headers: Union[Headers, Mapping[str, Union[str, bytes]], Iterable[Tuple[bytes, bytes]]] = ()
+    ) -> "Response":
+        """
+        Simplified API for creating response objects.
+        """
+        if isinstance(headers, Headers):
+            headers = headers
+        elif isinstance(headers, dict):
+            headers = Headers(
+                (always_bytes(k, "utf-8", "surrogateescape"),
+                 always_bytes(v, "utf-8", "surrogateescape"))
+                for k, v in headers.items()
+            )
+        elif isinstance(headers, Iterable):
+            headers = Headers(headers)
+        else:
+            raise TypeError("Expected headers to be an iterable or dict, but is {}.".format(
+                type(headers).__name__
+            ))
+
+        resp = cls(
+            b"HTTP/1.1",
+            status_code,
+            status_codes.RESPONSES.get(status_code, "").encode(),
+            headers,
+            None,
+            None,
+            time.time(),
+            time.time(),
+        )
+
+        # Assign this manually to update the content-length header.
+        if isinstance(content, bytes):
+            resp.content = content
+        elif isinstance(content, str):
+            resp.text = content
+        else:
+            raise TypeError(f"Expected content to be str or bytes, but is {type(content).__name__}.")
+
+        return resp
+
+    @property
+    def status_code(self) -> int:
+        """
+        HTTP Status Code, e.g. ``200``.
+        """
+        return self.data.status_code
+
+    @status_code.setter
+    def status_code(self, status_code: int) -> None:
+        self.data.status_code = status_code
+
+    @property
+    def reason(self) -> str:
+        """
+        HTTP Reason Phrase, e.g. "Not Found".
+        HTTP/2 responses do not contain a reason phrase, an empty string will be returned instead.
+        """
+        # Encoding: http://stackoverflow.com/a/16674906/934719
+        return self.data.reason.decode("ISO-8859-1")
+
+    @reason.setter
+    def reason(self, reason: Union[str, bytes]) -> None:
+        self.data.reason = strutils.always_bytes(reason, "ISO-8859-1")
+
+    def _get_cookies(self):
+        h = self.headers.get_all("set-cookie")
+        all_cookies = cookies.parse_set_cookie_headers(h)
+        return tuple(
+            (name, (value, attrs))
+            for name, value, attrs in all_cookies
+        )
+
+    def _set_cookies(self, value):
+        cookie_headers = []
+        for k, v in value:
+            header = cookies.format_set_cookie_header([(k, v[0], v[1])])
+            cookie_headers.append(header)
+        self.headers.set_all("set-cookie", cookie_headers)
+
+    @property
+    def cookies(self) -> multidict.MultiDictView:
+        """
+        The response cookies. A possibly empty
+        :py:class:`~mitmproxy.net.multidict.MultiDictView`, where the keys are cookie
+        name strings, and values are (value, attr) tuples. Value is a string,
+        and attr is an MultiDictView containing cookie attributes. Within
+        attrs, unary attributes (e.g. HTTPOnly) are indicated by a Null value.
+
+        Caveats:
+            Updating the attr
+        """
+        return multidict.MultiDictView(
+            self._get_cookies,
+            self._set_cookies
+        )
+
+    @cookies.setter
+    def cookies(self, value):
+        self._set_cookies(value)
+
+    def refresh(self, now=None):
+        """
+        This fairly complex and heuristic function refreshes a server
+        response for replay.
+
+            - It adjusts date, expires and last-modified headers.
+            - It adjusts cookie expiration.
+        """
+        if not now:
+            now = time.time()
+        delta = now - self.timestamp_start
+        refresh_headers = [
+            "date",
+            "expires",
+            "last-modified",
+        ]
+        for i in refresh_headers:
+            if i in self.headers:
+                d = parsedate_tz(self.headers[i])
+                if d:
+                    new = mktime_tz(d) + delta
+                    self.headers[i] = formatdate(new, usegmt=True)
+        c = []
+        for set_cookie_header in self.headers.get_all("set-cookie"):
+            try:
+                refreshed = cookies.refresh_set_cookie_header(set_cookie_header, delta)
+            except ValueError:
+                refreshed = set_cookie_header
+            c.append(refreshed)
+        if c:
+            self.headers.set_all("set-cookie", c)
 
 
 class HTTPFlow(flow.Flow):
@@ -16,8 +1110,8 @@ class HTTPFlow(flow.Flow):
     An HTTPFlow is a collection of objects representing a single HTTP
     transaction.
     """
-    request: http.Request
-    response: Optional[http.Response] = None
+    request: Request
+    response: Optional[Response] = None
     error: Optional[flow.Error] = None
     """
     Note that it's possible for a Flow to have both a response and an error
@@ -38,8 +1132,8 @@ class HTTPFlow(flow.Flow):
     _stateobject_attributes = flow.Flow._stateobject_attributes.copy()
     # mypy doesn't support update with kwargs
     _stateobject_attributes.update(dict(
-        request=http.Request,
-        response=http.Response,
+        request=Request,
+        response=Response,
         mode=str
     ))
 
@@ -67,8 +1161,8 @@ class HTTPFlow(flow.Flow):
 def make_error_response(
     status_code: int,
     message: str = "",
-    headers: Optional[http.Headers] = None,
-) -> http.Response:
+    headers: Optional[Headers] = None,
+) -> Response:
     body: bytes = """
         <html>
             <head>
@@ -86,18 +1180,18 @@ def make_error_response(
     ).encode("utf8", "replace")
 
     if not headers:
-        headers = http.Headers(
+        headers = Headers(
             Server=version.MITMPROXY,
             Connection="close",
             Content_Length=str(len(body)),
             Content_Type="text/html"
         )
 
-    return http.Response.make(status_code, body, headers)
+    return Response.make(status_code, body, headers)
 
 
-def make_connect_request(address: Tuple[str, int]) -> http.Request:
-    return http.Request(
+def make_connect_request(address: Tuple[str, int]) -> Request:
+    return Request(
         host=address[0],
         port=address[1],
         method=b"CONNECT",
@@ -105,7 +1199,7 @@ def make_connect_request(address: Tuple[str, int]) -> http.Request:
         authority=f"{address[0]}:{address[1]}".encode(),
         path=b"",
         http_version=b"HTTP/1.1",
-        headers=http.Headers(),
+        headers=Headers(),
         content=b"",
         trailers=None,
         timestamp_start=time.time(),
@@ -116,11 +1210,11 @@ def make_connect_request(address: Tuple[str, int]) -> http.Request:
 def make_connect_response(http_version):
     # Do not send any response headers as it breaks proxying non-80 ports on
     # Android emulators using the -http-proxy option.
-    return http.Response(
+    return Response(
         http_version,
         200,
         b"Connection established",
-        http.Headers(),
+        Headers(),
         b"",
         None,
         time.time(),
@@ -129,4 +1223,4 @@ def make_connect_response(http_version):
 
 
 def make_expect_continue_response():
-    return http.Response.make(100)
+    return Response.make(100)
