+++ mitmproxy/net/encoding.py
