--- mitmproxy/net/http/__init__.py
@@ -1,13 +1,0 @@
-from mitmproxy.net.http.request import Request
-from mitmproxy.net.http.response import Response
-from mitmproxy.net.http.message import Message
-from mitmproxy.net.http.headers import Headers, parse_content_type
-from mitmproxy.net.http import http1, status_codes, multipart
-
-__all__ = [
-    "Request",
-    "Response",
-    "Message",
-    "Headers", "parse_content_type",
-    "http1", "status_codes", "multipart",
-]
