+++ test/mitmproxy/net/http/test_response.py
@@ -3,8 +3,8 @@ import time
 import pytest
 from unittest import mock
 
-from mitmproxy.net.http import Headers
-from mitmproxy.net.http import Response
+from mitmproxy.http import Headers
+from mitmproxy.http import Response
 from mitmproxy.net.http.cookies import CookieAttrs
 from mitmproxy.test.tutils import tresp
 from .test_message import _test_passthrough_attr
