+++ test/mitmproxy/proxy/layers/http/test_http2.py
@@ -7,8 +7,8 @@ import pytest
 from h2.errors import ErrorCodes
 
 from mitmproxy.flow import Error
-from mitmproxy.http import HTTPFlow
-from mitmproxy.net.http import Headers, Request, status_codes
+from mitmproxy.http import HTTPFlow, Headers, Request
+from mitmproxy.net.http import status_codes
 from mitmproxy.proxy.layers.http import HTTPMode
 from mitmproxy.proxy.commands import CloseConnection, OpenConnection, SendData
 from mitmproxy.proxy.context import Context, Server
