+++ test/mitmproxy/proxy/layers/test_websocket.py
@@ -5,8 +5,7 @@ import pytest
 
 import wsproto
 import wsproto.events
-from mitmproxy.http import HTTPFlow
-from mitmproxy.net.http import Request, Response
+from mitmproxy.http import HTTPFlow, Request, Response
 from mitmproxy.proxy.layers.http import HTTPMode
 from mitmproxy.proxy.commands import SendData, CloseConnection, Log
 from mitmproxy.proxy.context import ConnectionState
